+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## Version 1.10.5
+## Version 1.11.0
 * Game Adjustments
+    * Angel Unbroken DR should now stack with all other DR
     * Destructive Dispel now calculates the DC based on the effective CL of the dispel and the highest mental stat to better support edge cases. Formula is 10 + 1/2 CL + Highest Mental Stat.
     * Dynamic item naming of armor no longer includes the enhancement bonus of the armor when it does not deviate from the original value
     * Expanded metamagic rod fixes to more unique rods
@@ -18,10 +19,12 @@
     * Combat Reflexes (Mythic)
     * Intimidating Prowess (Mythic)
     * Titan Strike
+    * Two-Weapon Defense (Mythic)
     * Warrior Priest (Mythic)
 * Added Feats
     * Dispel Focus
     * Greater Dispel Focus
+    * Two-Weapon Defense
 
 ## Version 1.10.4
 * Now works with 1.1.6e